requirement.txt
celery
pika
websocket-client
python-binance
pandas
numpy
pyyaml
